I edited the name and added programs from 8-10

+++ prog01.py
@@ -1,6 +1,6 @@
 # Create a program that ask user to input 2 numbers. Print the bigger number.
-num1 = int(input("Enter first number: "))
-num2 = int(input("Enter second number: "))
+num1 = float(input("Enter first number: "))
+num2 = float(input("Enter second number: "))
 
 if num1 > num2:
     print(num1)
